accdb: read chain_cnt from shmem in joiners instead of recomputing
The acc_map chain count determines the offset of every region in the
accdb shared memory layout, but the sizing formula was duplicated in
three files: fd_accdb_shmem.c (footprint/new), fd_accdb.c (fd_accdb_new
and fd_accdb_join_readonly) and fd_backup_cache.c
(fd_backup_cache_join).  Changing the formula in one place but not the
others makes a joiner silently compute wrong region offsets into live
shared memory -- in the best case an immediate SIGSEGV at tile init, in
the worst case silent corruption.

fd_accdb_shmem_new already stores the chain count it laid out in the
shmem header, so have every joiner read it back from there rather than
rederive it.  The formula now lives only in fd_accdb_shmem.c, next to
the layout it sizes.

+++ src/discof/backup/fd_backup_accidx.h
@@ -16,7 +16,7 @@ struct fd_backup_accidx {
   fd_accdb_accmeta_t const * acc_pool;     /* map ele pool */
   ulong                      max_accounts; /* map ele pool max */
   ulong                      seed;         /* map hash function */
-  uint                       chain_mask;   /* map chain count - 1 */
+  ulong                      chain_mask;   /* map chain count - 1 (chain counts exceed 2^32 at mainnet max_accounts) */
 
   ulong *       epoch_slot; /* epoch announced by this reader */
   ulong const * epoch;      /* accdb global epoch */
